frontend: ArticleTable: refactor moderators widget.
